refactor: extract session loading from AgentWorkspace
